Consolidate test modules
