+++ java-tests/validating/build.gradle.kts
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,29 +24,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        mavenLocal()
-    }
+import io.spine.dependency.lib.AutoService
+import io.spine.dependency.local.Logging
+import io.spine.dependency.local.Spine
+import io.spine.dependency.local.Validation
+
+plugins {
+    `java-test-fixtures`
 }
 
-rootProject.name = "validation"
+dependencies {
+    testFixturesAnnotationProcessor(AutoService.processor)
+    testFixturesCompileOnly(AutoService.annotations)
+    arrayOf(
+        Spine.base,
+        Logging.lib,
+        Validation.runtime
+    ).forEach {
+        testFixturesImplementation(it)
+    }
+    testImplementation(Spine.testlib)
+}
 
-include(
-    "proto",
-    ":proto:configuration",
-    ":proto:context",
-    "java",
-    "model",
-    "java-runtime",
-    "java-runtime-bundle",
-    "java-bundle",
-    ":java-tests",
-    ":java-tests:consumer",
-    ":java-tests:consumer-dependency",
-    ":java-tests:extensions",
-    ":java-tests:runtime",
-    ":java-tests:vanilla",
-    ":java-tests:validating",
-)
+testProtoDataRemoteDebug(enabled = false)
