+++ java-tests/validating/src/test/java/io/spine/test/protobuf/GeneratedMessageTest.java
