+++ java-tests/validating/src/test/java/io/spine/test/protobuf/package-info.java
