+++ java-tests/validating/src/test/java/io/spine/test/tools/validate/DistinctConstraintTest.java
