+++ java-tests/validating/src/test/java/io/spine/test/tools/validate/ExternalConstraintTest.java
