+++ java-tests/validating/src/test/java/io/spine/test/tools/validate/setonce/SetOnceAssertions.java
