+++ java-tests/validating/src/test/kotlin/io/spine/test/options/ValidatingOptionFactoryITest.kt
