+++ java-tests/validating/src/testFixtures/java/io/spine/test/options/BytesOptions.java
