+++ java-tests/validating/src/testFixtures/java/io/spine/test/options/package-info.java
