+++ java-tests/validating/src/testFixtures/java/io/spine/tools/validate/rule/AllRequired.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,29 +24,33 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        mavenLocal()
+package io.spine.tools.validate.rule;
+
+import com.google.errorprone.annotations.Immutable;
+import io.spine.code.proto.FieldContext;
+import io.spine.option.OptionsProto;
+import io.spine.validate.Constraint;
+import io.spine.validate.option.FieldValidatingOption;
+
+/**
+ * A field validating option which creates {@link AllRequiredConstraint}s.
+ *
+ * <p>The option reuses the {@link OptionsProto#required} extension.
+ */
+@Immutable
+public final class AllRequired extends FieldValidatingOption<Boolean> {
+
+    AllRequired() {
+        super(OptionsProto.required);
+    }
+
+    @Override
+    public boolean shouldValidate(FieldContext context) {
+        return context.targetDeclaration().isCollection() && super.shouldValidate(context);
+    }
+
+    @Override
+    public Constraint constraintFor(FieldContext field) {
+        return new AllRequiredConstraint(optionValue(field), field.targetDeclaration());
     }
 }
-
-rootProject.name = "validation"
-
-include(
-    "proto",
-    ":proto:configuration",
-    ":proto:context",
-    "java",
-    "model",
-    "java-runtime",
-    "java-runtime-bundle",
-    "java-bundle",
-    ":java-tests",
-    ":java-tests:consumer",
-    ":java-tests:consumer-dependency",
-    ":java-tests:extensions",
-    ":java-tests:runtime",
-    ":java-tests:vanilla",
-    ":java-tests:validating",
-)
