+++ java-tests/validating/src/testFixtures/proto/spine/test/options/bytes_direction_option.proto
