+++ java-tests/validating/src/testFixtures/proto/spine/test/protobuf/validating_builder_test.proto
