+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/custom_constraints.proto
