+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/entity.proto
