+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/external_constraint.proto
