+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/goes.proto
