+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/is_required.proto
