+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/name_clash.proto
