+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/required_field.proto
