+++ java-tests/validating/src/testFixtures/proto/spine/test/tools/validate/set_once.proto
