+++ java-tests/validating/src/testFixtures/proto/spine/test/validate/messages.proto
