+++ java-tests/validating/src/testFixtures/proto/spine/test/validate/patterns.proto
