+++ java-tests/validating/src/testFixtures/proto/spine/test/validate/validating_builder_test.proto
