--- java-tests/validation-gen/src/test/java/io/spine/test/tools/validate/rule/AllRequired.java
@@ -1,56 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.test.tools.validate.rule;
-
-import com.google.errorprone.annotations.Immutable;
-import io.spine.code.proto.FieldContext;
-import io.spine.option.OptionsProto;
-import io.spine.validate.Constraint;
-import io.spine.validate.option.FieldValidatingOption;
-
-/**
- * A field validating option which creates {@link AllRequiredConstraint}s.
- *
- * <p>The option reuses the {@link OptionsProto#required} extension.
- */
-@Immutable
-public final class AllRequired extends FieldValidatingOption<Boolean> {
-
-    AllRequired() {
-        super(OptionsProto.required);
-    }
-
-    @Override
-    public boolean shouldValidate(FieldContext context) {
-        return context.targetDeclaration().isCollection() && super.shouldValidate(context);
-    }
-
-    @Override
-    public Constraint constraintFor(FieldContext field) {
-        return new AllRequiredConstraint(optionValue(field), field.targetDeclaration());
-    }
-}
